requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12
Flask-Testing==0.6.2
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
numpy==1.12.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
Werkzeug==0.12.1
